can-daq-pc/requirements.txt
cantools
matplotlib
pyserial
customtkinter
numpy
tkcalendar